ResultView와 GamePlayView 분리
